Deal better with wrap-arounds in streaming.
Caught and fixed an edge case. Need to be careful when the current LL
entry is in a mini-LL that wraps around the end of LL memory.

Also, need to keep track of two LL entry indices at the left edge of LL
memory. This is because when we overwrite the beginning of LL memory,
the curAddr is related to the previous boundary LL entry. Once curAddr
has wrapped around the end, we can throw away the prior boundary LL
entry reference.

--BRJ and CAR