Eliminate enumerate_devices()
Seems superfluous with get_numDevices(), so remove a step. Also, it
seemed dangerous to remove an APS from the APSs map if it didn’t
respond to a broadcast. What if you have connected to a device not on
the local subnet and then called `enumerate_devices()`? This eliminates
that bug.